requirements.txt
flask==2.0.1
flask-jwt-extended==4.3.1
requests==2.20.0
Pillow==9.0.0
cryptography==3.3.1
pyjwt==1.7.1
pytest==7.1.2